+++ BrightWire/ExecutionGraph/Node/Layer/ElmanJordan.cs
@@ -89,7 +89,7 @@ namespace BrightWire.ExecutionGraph.Node.Layer
             var Uh = (FeedForward)_memory.FindByName("Uh");
 
             writer.Write(_isElman);
-            writer.Write(_inputSize);
+            writer.Write((int)_inputSize);
             writer.Write(_memory.Id);
             _memory.Data.WriteTo(writer);
             _Serialise(_activation, writer);
